// src/app/dashboard/location/[id]/staffs/schema.ts
import z from 'zod';

const InviteStaffSchema = z.object({
    firstName: z.string().min(2, { message: "Required" }),
    lastName: z.string().min(2, { message: "Required" }),
    phone: z.string().min(11, { message: 'Invalid phone number' }),
    email: z.string().min(8).email("invalid email."),
    role: z.string()
});

export {
    InviteStaffSchema
}